Allow running a single worker process (#31)
### Why?
Our daemon running setup today will either:
- Use the `SimpleRunner` (i.e. single process) if `--processes` is equal to 1.
- Use the `ProcessRunner` (i.e. one main process and `n` worker processes) is `--processes` is greater than 1. In that case, `--processes` is the number of worker processes.

This configuration setup doesn't allow running a main-worker setup with a single worker process. The daemon healthcheck implemented in https://github.com/globality-corp/microcosm-daemon/pull/24 relies on the master process to serve a `/health` endpoint, which means that in the current config setup we cannot have a single worker and a healthcheck. This is impractical e.g. when running services locally on Kubernetes, where we need a healthcheck endpoint but also want to reduce process count to reduce the memory footprint.

### What?
- Use a main+worker setup if the heartbeat config is turned on, i.e. if `--heartbeat-threshold-seconds` is greater than 0.
- Use the single-process setup if `--processes` is equal to 1 and the heartbeat config is off. Those are the default settings, so the default behavior remains unchanged.